+++ PerseusPluginLib/Basic/AddNoise.cs
@@ -43,12 +43,12 @@ namespace PerseusPluginLib.Basic{
 			}
 			foreach (int j in mainInds){
 				for (int i = 0; i < mdata.RowCount; i++){
-					mdata.Values.Set(i, j, mdata.Values.Get(i, j) + (float) rand.NextGaussian(0, std));
+					mdata.Values.Set(i, j, mdata.Values.Get(i, j) + rand.NextGaussian(0, std));
 				}
 			}
 			foreach (int j in numInds){
 				for (int i = 0; i < mdata.RowCount; i++){
-					mdata.NumericColumns[j][i] += (float) rand.NextGaussian(0, std);
+					mdata.NumericColumns[j][i] += rand.NextGaussian(0, std);
 				}
 			}
 		}
